requirements.txt
streamlit
google-genai
chromadb
pydantic
requests